tested reader.reducer

+++ src/redux/add.js
@@ -1,5 +1,5 @@
 const sum = (num1, num2) => {
-    return parseInt(num1) + parseInt(num2);
+    return parseFloat(num1) + parseFloat(num2);
 }
 
 export default sum;
